# tcl-web/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ToneCards Lab API",
    "description": "ToneCards Lab API for managing musical language flashcards",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cards/": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Create Card",
        "description": "Create a new card in a deck.",
        "operationId": "create_card_api_v1_cards__post",
        "parameters": [
          {
            "name": "deck_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Deck ID to add card to",
              "title": "Deck Id"
            },
            "description": "Deck ID to add card to"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Card_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/{card_id}": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "Get Card",
        "description": "Get a specific card by ID.",
        "operationId": "get_card_api_v1_cards__card_id__get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Card_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "cards"
        ],
        "summary": "Update Card",
        "operationId": "update_card_api_v1_cards__card_id__patch",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Card Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Card_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cards"
        ],
        "summary": "Delete Card",
        "description": "Delete a card from a deck.",
        "operationId": "delete_card_api_v1_cards__card_id__delete",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/deck/{deck_id}": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "Get Deck Cards",
        "description": "Get all cards in a deck, ordered by position.",
        "operationId": "get_deck_cards_api_v1_cards_deck__deck_id__get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_Card__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/deck/{deck_id}/reorder": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Reorder Cards",
        "description": "Reorder cards in a deck.",
        "operationId": "reorder_cards_api_v1_cards_deck__deck_id__reorder_post",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderCardsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/deck/{deck_id}/bulk-create": {
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Bulk Create Cards",
        "operationId": "bulk_create_cards_api_v1_cards_deck__deck_id__bulk_create_post",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CardCreate"
                },
                "title": "Cards Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_Card__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/deck/{deck_id}/all": {
      "delete": {
        "tags": [
          "cards"
        ],
        "summary": "Delete All Cards In Deck",
        "operationId": "delete_all_cards_in_deck_api_v1_cards_deck__deck_id__all_delete",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse"
      },
      "ApiResponse_Card_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Card"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse[Card]"
      },
...
      "ApiResponse_List_Card__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Card"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[Card]]"
      },
...
      "Card": {
        "properties": {
          "front_content": {
            "type": "string",
            "title": "Front Content",
            "description": "Content to display on front of card"
          },
          "back_content": {
            "type": "string",
            "title": "Back Content",
            "description": "Content to display on front of card"
          },
          "front_renderer": {
            "$ref": "#/components/schemas/RendererType",
            "description": "Renderer type for front content",
            "default": "string"
          },
          "back_renderer": {
            "$ref": "#/components/schemas/RendererType",
            "description": "Renderer type for back content",
            "default": "string"
          },
          "card_id": {
            "type": "string",
            "title": "Card Id",
            "description": "UUID of the card"
          }
        },
        "type": "object",
        "required": [
          "front_content",
          "back_content",
          "card_id"
        ],
        "title": "Card",
        "description": "Full card model with ID."
      },
      "CardCreate": {
        "properties": {
          "front_content": {
            "type": "string",
            "title": "Front Content",
            "description": "Content to display on front of card"
          },
          "back_content": {
            "type": "string",
            "title": "Back Content",
            "description": "Content to display on front of card"
          },
          "front_renderer": {
            "$ref": "#/components/schemas/RendererType",
            "description": "Renderer type for front content",
            "default": "string"
          },
          "back_renderer": {
            "$ref": "#/components/schemas/RendererType",
            "description": "Renderer type for back content",
            "default": "string"
          }
        },
        "type": "object",
        "required": [
          "front_content",
          "back_content"
        ],
        "title": "CardCreate",
        "description": "Request to create a new card."
      },
      "CardUpdate": {
        "properties": {
          "front_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Front Content",
            "description": "Updated front content"
          },
          "back_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Back Content",
            "description": "Updated back content"
          },
          "front_renderer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RendererType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Updated front renderer"
          },
          "back_renderer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RendererType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Updated back renderer"
          }
        },
        "type": "object",
        "title": "CardUpdate",
        "description": "Request to update an existing card."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Metadata": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Response timestamp"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique request identifier"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "API version",
            "default": "1.0"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count",
            "description": "Total number of items (for paginated responses)"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page",
            "description": "Current page number"
          },
          "page_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Size",
            "description": "Number of items per page"
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms",
            "description": "Request execution time in milliseconds"
          }
        },
        "type": "object",
        "title": "Metadata"
      },
      "RendererType": {
        "type": "string",
        "enum": [
          "string",
          "markdown",
          "abc_js",
          "image",
          "latex",
          "mermaid"
        ],
        "title": "RendererType",
        "description": "Available content renderer types."
      },
      "ReorderCardsRequest": {
        "properties": {
          "card_order": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Card Order",
            "description": "Ordered list of card IDs representing new order"
          }
        },
        "type": "object",
        "required": [
          "card_order"
        ],
        "title": "ReorderCardsRequest",
        "description": "Request to reorder cards in a deck."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}